+++ config/default.js
@@ -19,6 +19,7 @@ module.exports = {
       '@elifesciences/component-model-team',
       '@elifesciences/component-model-user',
       '@elifesciences/component-model',
+      '@elifesciences/component-services',
       '@elifesciences/xpub-server',
       '@elifesciences/component-dashboard',
       '@elifesciences/component-submission',
