+++ packages/component-services/index.js
@@ -1,0 +1,5 @@
+const S3Storage = require('./server/s3Storage')
+
+module.exports = {
+  S3Storage,
+}
