+++ packages/component-services/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@elifesciences/component-services",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT"
+}
