+++ packages/component-submission/server/resolvers/submitManuscript.test.js
@@ -10,7 +10,7 @@ const mailer = require('@pubsweet/component-send-email')
 const { mecaExport } = require('@elifesciences/component-meca')
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
-const { S3Storage } = require('@elifesciences/xpub-client')
+const { S3Storage } = require('@elifesciences/component-services')
 
 const { Mutation } = require('.')
 const {
