+++ packages/component-submission/server/resolvers/uploadSupportingFile.js
@@ -1,6 +1,6 @@
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
-const { S3Storage } = require('@elifesciences/xpub-client')
+const { S3Storage } = require('@elifesciences/component-services')
 
 const { SupportingFiles } = require('../use-cases')
