+++ scripts/mecaExport.js
@@ -1,6 +1,6 @@
 const { db } = require('pubsweet-server')
 const { mecaExport } = require('@elifesciences/component-meca')
-const { S3Storage } = require('@elifesciences/xpub-client')
+const { S3Storage } = require('@elifesciences/component-services')
 const Manuscript = require('@elifesciences/component-model-manuscript').model
 
 const doExport = async (id, getContent, ip) => {
