+++ server/xpub-client/index.js
@@ -1,7 +1,1 @@
-const S3Storage = require('./s3Storage')
-const ScienceBeamApi = require('./scienceBeamApi')
-
-module.exports = {
-  S3Storage,
-  ScienceBeamApi,
-}
+module.exports = {}
